chore: move checkProjectsManifest to find-workspace-packages
