+++ workspace/find-workspace-packages/test/__fixtures__/warning-for-non-root-project/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "root",
+  "version": "1.0.0",
+  "pnpm": {
+    "overrides": {}
+  }
+}
