+++ workspace/find-workspace-packages/test/__fixtures__/warning-for-non-root-project/packages/bar/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "bar",
+  "version": "1.0.0",
+  "pnpm": {},
+  "resolutions": {}
+}
