+++ workspace/find-workspace-packages/test/__fixtures__/warning-for-non-root-project/packages/foo/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "foo",
+  "version": "1.0.0",
+  "pnpm": {}
+}
