+++ workspace/find-workspace-packages/test/__fixtures__/warning-for-non-root-project/pnpm-workspace.yaml
@@ -1,0 +1,2 @@
+packages:
+  - packages/*
