+++ workspace/find-workspace-packages/test/index.ts
@@ -1,5 +1,14 @@
 import path from 'path'
-import { findWorkspacePackagesNoCheck, arrayOfWorkspacePackagesToMap } from '@pnpm/find-workspace-packages'
+import { findWorkspacePackagesNoCheck, arrayOfWorkspacePackagesToMap, findWorkspacePackages } from '@pnpm/find-workspace-packages'
+import { logger } from '@pnpm/logger'
+
+beforeEach(() => {
+  jest.spyOn(logger, 'warn')
+})
+
+afterEach(() => {
+  (logger.warn as jest.Mock).mockRestore()
+})
 
 // This is supported for compatibility with Yarn's implementation
 // see https://github.com/pnpm/pnpm/issues/2648
@@ -20,4 +29,17 @@ test('findWorkspacePackagesNoCheck() skips engine checks', async () => {
   const pkgs = await findWorkspacePackagesNoCheck(path.join(__dirname, '__fixtures__/bad-engine'))
   expect(pkgs.length).toBe(1)
   expect(pkgs[0].manifest.name).toBe('pkg')
+})
+
+test('findWorkspacePackages() output warnings for non-root workspace project', async () => {
+  const fixturePath = path.join(__dirname, '__fixtures__/warning-for-non-root-project')
+
+  const pkgs = await findWorkspacePackages(fixturePath)
+  expect(pkgs.length).toBe(3)
+  expect(logger.warn).toBeCalledTimes(3)
+  const fooPath = path.join(fixturePath, 'packages/foo')
+  const barPath = path.join(fixturePath, 'packages/bar')
+  expect(logger.warn).toHaveBeenNthCalledWith(1, { prefix: barPath, message: `"pnpm" was found in ${barPath}/package.json. It will not take effect, should configure "pnpm" at the root of the project.` })
+  expect(logger.warn).toHaveBeenNthCalledWith(2, { prefix: barPath, message: `"resolutions" was found in ${barPath}/package.json. It will not take effect, should configure "resolutions" at the root of the project.` })
+  expect(logger.warn).toHaveBeenNthCalledWith(3, { prefix: fooPath, message: `"pnpm" was found in ${fooPath}/package.json. It will not take effect, should configure "pnpm" at the root of the project.` })
 })
